[FEATURE] Phar building
Brings the ability to package surf as PHP archive file (phar).
This is done with the help of the "box" project (https://github.com/box-project/box2)

Note: This disables the the Dotenv reading as we cannot guarantee there is a .env file
in __DIR__ and relative paths in a phar are tricky.

# Building

```
# Install box globally (remember to add ~/.composer/vendor/bin to $PATH)
~$ curl -LSs http://box-project.org/installer.php | php

# Build
~$ box build

#Test
~$ php surf.phar
```